storage_research/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.0
exceptiongroup==1.1.2
fonttools==4.41.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psycopg2-binary==2.9.6
pydantic==2.0a3
pydantic-settings==1.99
pydantic_core==0.25.0
pymongo==4.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
typing_extensions==4.7.1
tzdata==2023.3
